Fix a bug in webgl storage and fix demos (#505)
- Fix a bug in webgl where the user can't directly interact with the backend like in `imagenet` and `cppn` demos (e.g. calling `backend.getTexture(id)` right after calling `NDArray.make()`)
- Make the backend storage mehanism clearer (separate registration from writing).
- Small changes of hyper-params in the complementary colors demo, to converge faster
- To be more compatible with TF, `add/sub/multiply/divide` now require both inputs to be of the same dtype. Update outdated demos and squeezenet model.

After this PR is in, I will:
- release a new `deeplearn` package
- release a new `deeplearn-squeezenet` package
- update `demos/package.json` to depend on the newest libraries

+++ models/squeezenet/squeezenet.ts
@@ -67,7 +67,8 @@ export class SqueezeNet implements Model {
       let activation: Array3D;
       // Preprocess the input.
       const preprocessedInput =
-          this.math.subtract(input, this.preprocessOffset) as Array3D;
+          this.math.subtract(input.asType('float32'), this.preprocessOffset) as
+          Array3D;
       const conv1 = this.math.conv2d(
           preprocessedInput, this.variables['conv1_W:0'] as Array4D,
           this.variables['conv1_b:0'] as Array1D, 2, 0);
